refactor navigation parsing to support `@NavDestination`
* Replace `Parser.kt` with `NavDestinationParser.kt` and consolidate parsing logic
* Introduce `parseNavDestinationData` to handle new `SCREEN`, `OVERLAY`, and `TAB_ROOT` destination types
* Implement `parseScreenLike` to handle stack screens and modal overlays, supporting optional `@AssistedFactory` wiring
* Implement `parseTabLike` for tab root destinations, enforcing plain `@Inject` usage without assisted parameters
* Update validation logic to ensure parameterized presenters have exactly one `@Assisted` constructor parameter
* Remove legacy parsing logic for `@NavScreen`, `@TabScreen`, and `@NavSheet` annotations
* Improve diagnostic logging and KDoc documentation for the parsing process
